Support v128 const load store and extract lane in bbq only with O0 register allocation https://bugs.webkit.org/show_bug.cgi?id=245897 <rdar://problem/100772941> https://github.com/WebKit/WebKit/pull/4879
Reviewed by: Nobody (OOPS).

This is the first patch to support the WASM SIMD proposal:
https://github.com/WebAssembly/simd/blob/main/proposals/simd/SIMD.md

- Add a feature flag and support for parsing the new SIMD opcodes and the V128 type
When this flag is enabled, we disable LLInt and OMG, as well as the Linear Scan
and Graph Coloring register allocators (for now).

- Implement load, store, extract_lane and v128.const on both Intel and ARM64.
This is meant to be enough instruction implementation to test that the other parts
of this patch work.

- Spill and fill 128-bit vectors when they are used

Next steps
- Every place where we see an Options::useWebAssemblySIMD() ? should eventually be removed.
These were added for now so that the patch would be perf-neutral on JS2, but
the register allocator should eventually be able to see that if the full width is
never used, we should only spill a double.